requirements.txt



# Requirements pour le Document Processor
# Phase 1-3: Configuration, Téléchargement, Extraction PDF

# Téléchargement et requêtes HTTP
requests>=2.31.0
urllib3>=2.0.0

# Extraction PDF - Méthodes multiples
PyPDF2>=3.0.1
pdfplumber>=0.9.0
pdfminer.six>=20211012

# Nettoyage et traitement de texte
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Machine Learning et NLP (pour phases futures)
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0

# Configuration et variables d'environnement
python-dotenv>=1.0.0

# Utilitaires
pathlib2>=2.3.0  # Pour compatibilité Python < 3.6
unicodedata2>=15.0.0  # Normalisation Unicode avancée

# Tests et développement
pytest>=7.4.0
pytest-cov>=4.1.0
mock>=5.1.0

# Logging et debugging
colorlog>=6.7.0

# APIs externes (pour phases futures)
arxiv>=1.4.0  # API arXiv officielle
feedparser>=6.0.0  # Parsing RSS/Atom

# Sérialisation et formats
jsonschema>=4.18.0
pyyaml>=6.0.0

# Validation et sécurité
validators>=0.20.0
certifi>=2023.7.0

# Dépendances optionnelles pour des fonctionnalités étendues
# pandas>=1.3.0  # Pour traitement des données
# numpy>=1.21.0  # Pour calculs numériques
# matplotlib>=3.4.0  # Pour visualisations
# Pillow>=8.3.0  # Pour traitement d'images